requirements.txt
natsort==8.4.0
ncls==0.0.68
numpy==1.26.2
pandas==2.1.4
prettytable==3.9.0
pyranges==0.0.129
python-dateutil==2.8.2
pytz==2023.3.post1
PyVCF3==1.0.3
six==1.16.0
sorted-nearest==0.0.39
tabulate==0.9.0
tzdata==2023.3
wcwidth==0.2.12
